moved transcript function to new class to follow SRP
